fix(cli): stop walking the Windows registry on every non-winget startup
`aspire run`, `aspire add`, every CLI command on Windows currently walks
both HKCU and HKLM `Uninstall` hives looking for the winget portable
manifest entry, on every single startup, for any install path that
didn't ship the `.aspire-install.json` sidecar (notably dev builds and
legacy pre-sidecar installs). The probe has a positive fast path —
sidecar exists, skip — but no negative-result cache. The old
`{"source":"winget"}` byte-array writer only fires when the registry
confirms a winget install, so non-winget Windows users re-walk the
registry forever. The factory comment in `Program.cs` claimed "the
worst case here is a no-op", which is true about side effects but
misleading about cost.

Rework the probe as a sidecar back-fill that always writes after one
registry read, either the positive `{"source":"winget"}` payload (when
the running binary IS a winget portable install) or a
`{"backfilled":true}` sentinel otherwise. The sentinel has no `source`
field, so `InstallSidecarReader.TryRead` parses it as Unknown and
downstream consumers see the row identically to a missing sidecar. The
existing `File.Exists` fast path at the top of `EnsureSidecar` then
short-circuits every subsequent startup — `HasWingetAspireUninstallEntry`
runs at most once per install location, ever.

Rename for clarity:

- `WingetFirstRunProbe` → `WingetSidecarBackfill`. "FirstRun" implied
  run-once-and-never-again; the reality is "every run until a sidecar
  exists". "Probe" implied read-only detection; the type writes.
- `WingetFirstRunProbe.Run(string)` → `WingetSidecarBackfill.EnsureSidecar(string)`.
  The new name matches the post-call invariant: a sidecar exists at the
  binary directory (real or backfilled sentinel) unless the write itself
  failed.
- `Program.TryRunWingetFirstRunProbe` → `Program.TryEnsureWingetSidecar`.
- Test files `WingetRegistryProbeTests` / `WingetStartupProbeTests` →
  `WingetSidecarBackfillTests` / `WingetSidecarBackfillStartupTests`.

Surprises worth calling out:

- One small behavior change for non-winget Windows installs: discovery
  used to mark these rows as `notProbed` (no sidecar = "we won't execute
  this binary"). Post-change the sidecar exists (sentinel form), so
  discovery promotes the row to `ok` and spawns the peer probe. The
  spawn target is the running binary itself — same binary that just
  wrote the sentinel — so self-vouching is logically consistent; the
  `notProbed` outcome was a side effect of the missing-sidecar gate,
  not a deliberate categorization. End user effect: the running CLI's
  row in `aspire --info` shows a real version instead of "not probed".
- A user who later installs via winget on top of a previously
  probed-negative location pays one row of `source: null` until they
  reinstall or delete the sentinel. `EnsureSidecar_DoesNotPromoteBackfilledSentinel_OnLaterWingetRun`
  pins this trade-off explicitly. In practice winget portable installs
  land in their own directory under `%LOCALAPPDATA%\Microsoft\WinGet\Packages\`,
  so the collision is rare.
- The startup-contract tests (`WingetSidecarBackfillStartupTests`) now
  clean up the sentinel written next to the test runner binary; without
  cleanup, the second test would skip via the existing-sidecar fast path
  and pollute the build output across runs.

Tests:

- `WingetSidecarBackfillTests`:
  - `EnsureSidecar_WritesWingetSidecar_WhenRegistryClaimsWinget` — positive path.
  - `EnsureSidecar_WritesBackfilledSentinel_WhenRegistryDoesNotClaimWinget` — sentinel content + asserts `source` field intentionally absent.
  - `EnsureSidecar_SkipsRegistryWalk_WhenSidecarExists` — counting fake registry reader, asserts `CallCount == 0` (the cost-reduction guarantee).
  - `EnsureSidecar_DoesNotOverwriteExistingSidecar` — idempotency.
  - `EnsureSidecar_DoesNotPromoteBackfilledSentinel_OnLaterWingetRun` — trade-off pinned.
  - `EnsureSidecar_ConcurrentInvocations_ProduceSingleValidSidecar` `[Theory]` — both positive and negative concurrent paths.
- `WingetSidecarBackfillStartupTests`: renamed + cleanup of negative-result sentinel.
